[lineuparr] Bump to 1.26.1791747 (shared matcher core); add matching_core.py
Lays down the shared matcher core migration. ADDS matching_core.py: the Hub
folder predated the migration, and fuzzy_matcher.py imports FuzzyMatcherCore
from it, so the packaged plugin will not import without it.

Changed: fuzzy_matcher.py (now subclasses FuzzyMatcherCore), plugin.py and
plugin.json (version 1.26.1641222 to 1.26.1791747), plus new matching_core.py
(vendored byte-identically from the shared core). Other files unchanged.
