+++ plugins/lineuparr/fuzzy_matcher.py
@@ -10,6 +10,25 @@ import re
 import logging
 import unicodedata
 
+# The pure matching primitives (normalize_name, calculate_similarity, the callsign
+# ladder, ...) live in the vendored shared core. The rapidfuzz fast path and its
+# pure-Python fallback are inside FuzzyMatcherCore.calculate_similarity. The decorative
+# helpers are re-exported so tests/callers that reference them keep working.
+try:
+    from .matching_core import (
+        FuzzyMatcherCore,
+        _is_decorative_char,
+        _normalize_emoji,
+        _strip_stylized_tokens,
+    )
+except ImportError:  # script/test context without the package parent on sys.path
+    from matching_core import (
+        FuzzyMatcherCore,
+        _is_decorative_char,
+        _normalize_emoji,
+        _strip_stylized_tokens,
+    )
+
 __version__ = "1.3.4"
 
 LOGGER = logging.getLogger("plugins.lineuparr.fuzzy_matcher")
@@ -19,37 +38,12 @@ if not LOGGER.handlers:
     LOGGER.addHandler(_handler)
 LOGGER.setLevel(logging.DEBUG)
 
-# --- Pattern categories for normalization ---
-
-# Unicode categories considered decorative/badge characters by IPTV providers.
-# So = Other Symbol (◉), No = Other Number (², ³), Lm = Modifier Letter
-# (ᴿᴬᵂ, ᴴᴰ, ⱽᴵᴾ superscripts), Sk = Modifier Symbol.
-# Accented letters (é, î, ü…) are Ll/Lu and are NOT in this set.
-# Sm (Math Symbol) is intentionally EXCLUDED: it contains "+", which is a
-# meaningful, channel-distinguishing character (Canal+, Three Stooges+,
-# Comedy Central+). Stripping it regresses those matches.
-_DECORATOR_CATS = frozenset({'So', 'No', 'Lm', 'Sk'})
-
-# Tokens that are non-distinctive stream-label variants (e.g. "ABC News Live"
-# should still match "ABC News"). Used by the subset/divergent guards.
-_NON_DISTINCTIVE_TOKENS = frozenset({"live", "now", "new"})
-
-def _is_distinctive(t):
-    """Return True if token t is distinctive enough to matter in subset/divergent guards."""
-    return t not in _NON_DISTINCTIVE_TOKENS and (len(t) >= 4 or (t.isdigit() and len(t) >= 2))
 
 # Matches "+1" / "+2" time-shift suffixes in the ORIGINAL (pre-normalization)
 # channel name. Must be checked before normalization because "+" is in the Sm
 # category and gets stripped, making "+1" indistinguishable from "1" afterward.
 _PLUS_SHIFT_RE = re.compile(r'\+\s{0,2}\d{1,2}\b')
 
-QUALITY_PATTERNS = [
-    r'\s*\[(4K|8K|UHD|FHD|HD|HDR|HEVC|SD|FD|Unknown|Unk|Slow|Dead|Backup)\]\s*',
-    r'\s*\((4K|8K|UHD|FHD|HD|HDR|HEVC|SD|FD|Unknown|Unk|Slow|Dead|Backup)\)\s*',
-    r'^\s*(4K|8K|UHD|FHD|HD|HDR|HEVC|SD|FD|Unknown|Unk|Slow|Dead)\b\s*',
-    r'\s*\b(4K|8K|UHD|FHD|HD|HDR|HEVC|SD|FD|Unknown|Unk|Slow|Dead)$',
-    r'\s+\b(4K|8K|UHD|FHD|HD|HDR|HEVC|SD|FD|Unknown|Unk|Slow|Dead)\b\s+',
-]
 
 # Quality markers that distinguish an upgrade-tier channel from its standard twin.
 # HD/FHD/HEVC are intentionally excluded - they don't create a separate twin channel.
@@ -61,60 +55,26 @@ def has_upgrade_quality(name: str) -> bool:
     return bool(_UPGRADE_QUALITY_RE.search(name))
 
 
-REGIONAL_PATTERNS = [
-    # East/West are intentionally NOT stripped - they distinguish separate channel feeds
-    # (e.g., "HBO East" and "HBO West" are different channels)
-    r'\s[Pp][Aa][Cc][Ii][Ff][Ii][Cc]',
-    r'\s[Cc][Ee][Nn][Tt][Rr][Aa][Ll]',
-    r'\s[Mm][Oo][Uu][Nn][Tt][Aa][Ii][Nn]',
-    r'\s[Aa][Tt][Ll][Aa][Nn][Tt][Ii][Cc]',
-    r'\s*\([Pp][Aa][Cc][Ii][Ff][Ii][Cc]\)\s*',
-    r'\s*\([Cc][Ee][Nn][Tt][Rr][Aa][Ll]\)\s*',
-    r'\s*\([Mm][Oo][Uu][Nn][Tt][Aa][Ii][Nn]\)\s*',
-    r'\s*\([Aa][Tt][Ll][Aa][Nn][Tt][Ii][Cc]\)\s*',
-]
+# Country tokens for the delimited provider-prefix patterns below; curated so a
+# bare delimited word ("(SPORTS)") isn't misread as a country. Keep in sync with
+# detect_stream_country().
+_PREFIX_COUNTRY = r'US|USA|UK|CA|AU|FR|DE|ES|IT|NL|BR|MX|IN'
 
-GEOGRAPHIC_PATTERNS = [
-    r'\b[A-Z]{2,3}:\s*',
-    r'\b[A-Z]{2,3}\s*-\s*',
-    r'\|[A-Z]{2,3}\|\s*',
-    r'\[[A-Z]{2,3}\]\s*',
-]
+# (open, close) pairs; open and close must MATCH, so "(US]" / "│US)" are rejected.
+_DELIM_PAIRS = ((r'\(', r'\)'), (r'\[', r'\]'), (r'\|', r'\|'), ('┃', '┃'), ('│', '│'))
 
-# Enhanced provider prefix patterns for IPTV-specific naming
-PROVIDER_PREFIX_PATTERNS = [
-    r'^(?:US|USA|UK|CA|AU|FR|DE|ES|IT|NL|BR|MX|IN)\s*[:\-\|]\s*',
-    # Bare country tag + whitespace, no separator (e.g. "US Racer Network",
-    # "FR beIN SPORTS MAX", "MEX Bein Sports"). Restricted to a curated set so
-    # it cannot eat a real channel name: "USA Network" (USA != US + space),
-    # "In Country Television" ("IN") and "IT Crowd" ("IT") are all safe too.
-    # Keep this set in sync with detect_stream_country()'s bare-space branch.
-    r'^(?:US|UK|CA|AU|FR|DE|MX|MEX|FRA|GER)\s+',
-    # "USA " space prefix as a US country tag ("USA  ABC", "USA BET"). A
-    # negative lookahead for NETWORK protects the real channel "USA Network"
-    # (these feeds tag that one as "US ..."/"US: ...", never bare "USA ").
-    # Keep in sync with detect_stream_country()'s USA branch.
-    r'^USA\s+(?!NETWORK\b)',
-    # Country code glued directly to a quality tag with no separator
-    # ("UKSD: Sky Sports", "UKHD ESPN", "USFHD ..."). Detection mirrors this
-    # in detect_stream_country() so these can't leak as wildcards.
-    r'^(?:US|UK)(?:SD|HD|FHD|UHD|FD|HEVC|4K|8K)\b\s*[:\-\|]?\s*',
-    r'^\s*\((?:US|USA|UK|CA|AU|FR|DE|ES|IT|NL|BR|MX|IN)\)\s*',
-    r'\s*\|\s*(?:US|USA|UK|CA|AU|FR|DE|ES|IT|NL|BR|MX|IN)\s*$',
-    # Content-category group prefixes used by some IPTV providers.
-    r'^(?:ADULT|EROTIC|PRIME|GOLD)\s*[:\-\|]\s*',
-    # FAST streaming-platform source tags (Roku, Tubi, Pluto, etc.). These mark
-    # the distribution platform, not the channel or its country, so strip them
-    # for matching ("RK: beIN Sports Xtra" -> "beIN Sports Xtra"). A separator
-    # is required so this can't eat real names like "GOLF" or "PLEX TV Movies".
-    # NOT a country signal: detect_stream_country() ignores these (correctly,
-    # since a platform like Roku spans US/CA/UK).
-    r'^(?:RK|GO|TUBI|PLUTO|XUMO|PLEX|STIRR|FREEVEE|GLANCE)\s*[:\-\|]\s*',
-]
 
-MISC_PATTERNS = [
-    r'\s*\([^)]*\)\s*',
-]
+def _balanced_delim(token):
+    """Regex fragment matching `token` wrapped in one MATCHED delimiter pair:
+    (token) [token] |token| ┃token┃ │token│. `token` may contain a capture group."""
+    return '(?:' + '|'.join(
+        o + r'\s*(?:' + token + r')\s*' + c for o, c in _DELIM_PAIRS
+    ) + ')'
+
+
+# Leading country tag in a matched delimiter pair; one capture group fires.
+_BRACKETED_CC_RE = re.compile(r'^\s*' + _balanced_delim(r'([A-Za-z]{2,3})'))
+
 
 # ISO-2 country codes Lineuparr lineup filenames use (keep in sync with
 # PROVIDER_PREFIX_PATTERNS above and PluginConfig.COUNTRY_DIR_MAP).
@@ -226,11 +186,12 @@ def detect_stream_country(name):
         mapped = _PLUTO_COUNTRY_MAP.get(m.group(1).upper())
         return mapped if mapped in _KNOWN_COUNTRY_CODES else None
 
-    m = re.match(r'^\s*\(\s*([A-Za-z]{2,3})\s*\)', name)
+    m = _BRACKETED_CC_RE.match(name)
     if m:
-        return _normalize_country_token(m.group(1))
+        tok = next((g for g in m.groups() if g), None)
+        return _normalize_country_token(tok) if tok else None
 
-    m = re.match(r'^\s*([A-Za-z]{2,3})\s*[-:|]', name)
+    m = re.match(r'^\s*([A-Za-z]{2,3})\s*[-:|┃│]', name)
     if m:
         return _normalize_country_token(m.group(1))
 
@@ -295,7 +256,7 @@ def detect_category_country(category_name):
         return None
 
     # Country code followed by a delimiter: "AU| ...", "UK: ...", "US-...".
-    m = re.match(r'^\s*([A-Za-z]{2,3})\s*[-:|]', category_name)
+    m = re.match(r'^\s*([A-Za-z]{2,3})\s*[-:|┃│]', category_name)
     if m:
         return _normalize_country_token(m.group(1))
 
@@ -329,22 +290,23 @@ def country_codes_in_text(text):
     return codes
 
 
-class FuzzyMatcher:
+class FuzzyMatcher(FuzzyMatcherCore):
     """Handles fuzzy matching for Lineuparr with alias support and channel number boosting."""
 
+    # Lineuparr strips bare time-zone region words (Pacific/Central/Mountain/Atlantic) for
+    # SCORING and enforces region correctness via its post-match region filter
+    # (match_all_streams reads the ORIGINAL names). See REGIONAL_BARE_PATTERNS in the core.
+    _STRIP_BARE_REGION = True
+
     def __init__(self, match_threshold=80, logger=None):
-        self.match_threshold = match_threshold
-        self.logger = logger or LOGGER
-        # Cache for pre-normalized stream names (performance optimization)
-        self._norm_cache = {}  # raw_name -> normalized_lower
-        self._norm_nospace_cache = {}  # raw_name -> normalized_nospace
-        self._processed_cache = {}  # raw_name -> processed_for_matching
+        # The core seeds match_threshold, logger, the four normalization/callsign
+        # caches, and the known-callsign rescue slot (unused here - Lineuparr has no
+        # channel DB). Only the country-filter counter is Lineuparr-specific.
+        super().__init__(match_threshold=match_threshold, logger=logger or LOGGER)
         # Cumulative count of candidates dropped by the country filter across
         # match_all_streams calls. Callers reset this before a matching loop
         # and read it after, so they can log a summary.
         self.country_filter_drops = 0
-        # Cache for callsign extraction: raw_name -> (callsign|None, is_high_conf)
-        self._callsign_cache = {}
 
     def precompute_normalizations(self, names, user_ignored_tags=None):
         """
@@ -388,300 +350,6 @@ class FuzzyMatcher:
             return None
         return self.process_string_for_matching(norm)
 
-    def normalize_name(self, name, user_ignored_tags=None, ignore_quality=True, ignore_regional=True,
-                       ignore_geographic=True, ignore_misc=True):
-        """
-        Normalize channel or stream name for matching by removing tags, prefixes, and noise.
-        """
-        if user_ignored_tags is None:
-            user_ignored_tags = []
-
-        original_name = name
-
-        # Strip IPTV provider prefixes BEFORE hyphen normalization so that
-        # "FR - Canal+ FHD" loses its "FR - " while the hyphen is still a
-        # hyphen. After hyphen normalization the separator would become a space
-        # and the pattern would fail to match, leaving "FR" as a stray token.
-        for pattern in PROVIDER_PREFIX_PATTERNS:
-            name = re.sub(pattern, '', name, flags=re.IGNORECASE)
-
-        # Quality patterns (before space normalization). Loop until stable so
-        # chained suffixes like "4K HDR" or "UHD HDR" are fully stripped in
-        # successive passes (each pass may expose a token for the next).
-        if ignore_quality:
-            prev = None
-            while prev != name:
-                prev = name
-                for pattern in QUALITY_PATTERNS:
-                    name = re.sub(pattern, '', name, flags=re.IGNORECASE)
-
-        # Normalize spacing around numbers
-        name = re.sub(r'([a-zA-Z])(\d)', r'\1 \2', name)
-        name = re.sub(r'(\d)([a-zA-Z])', r'\1 \2', name)
-
-        # Normalize hyphens to spaces
-        name = re.sub(r'-', ' ', name)
-
-        # Replace dots between word chars with spaces (e.g. "JusticeCentral.TV"
-        # → "JusticeCentral TV"). Keeps the dot-suffix variant equivalent to
-        # the spaced form for matching purposes.
-        name = re.sub(r'(?<=\w)\.(?=\w)', ' ', name)
-
-        # Normalize number-words to digits so "BBC Three" and "BBC 3" share
-        # tokens. Critical for cases like "Three Angels Broadcasting Network"
-        # vs "3 Angels Broadcasting Network", and for BBC One/Two/Three/Four
-        # vs BBC 1/2/3/4. Bounded by word boundaries so brand names with
-        # embedded letters (e.g. "Onesimus") aren't corrupted.
-        _NUM_WORDS = {
-            "one": "1", "two": "2", "three": "3", "four": "4", "five": "5",
-            "six": "6", "seven": "7", "eight": "8", "nine": "9", "ten": "10",
-            "eleven": "11", "twelve": "12",
-        }
-        def _num_repl(m):
-            return _NUM_WORDS[m.group(0).lower()]
-        name = re.sub(
-            r'\b(one|two|three|four|five|six|seven|eight|nine|ten|eleven|twelve)\b',
-            _num_repl, name, flags=re.IGNORECASE,
-        )
-
-        # Split CamelCase boundaries so "JusticeCentral" becomes "Justice
-        # Central" and "DangerTV" becomes "Danger TV". Two separate patterns:
-        #   1. lower → Upper followed by lower  (Justice|Central, Danger|Iq → no, etc.)
-        #   2. lower(4+) → UPPER acronym at word boundary  (Danger|TV, Beauty|IQ)
-        # The 4-char floor on rule 2 protects short brand names like "MeTV" and
-        # "truTV" whose existing EPG matches rely on the un-split form.
-        name = re.sub(r'([a-z])([A-Z][a-z])', r'\1 \2', name)
-        name = re.sub(r'([a-z]{4,})([A-Z]{2,})\b', r'\1 \2', name)
-
-        # Strip region tokens (East / Eastern / West and the (E)/(W)
-        # abbreviations) for SCORING. Region CORRECTNESS - East vs West vs
-        # Pacific - is enforced separately by the post-match region filter in
-        # match_all_streams, which reads the ORIGINAL un-normalized names, not
-        # this output. Stripping here lets a regionless lineup channel
-        # ("Food Network") still score-match a region-tagged stream
-        # ("Food Network Eastern" / "... East") instead of being dragged below
-        # threshold by the extra token; the filter then accepts it (regionless
-        # defaults to East). "Western"/"Westerns" is a movie genre, NOT a
-        # region, so bare "west" is stripped only on a word boundary (\bwest\b
-        # does not match inside "western") and "western" is never touched.
-        name = re.sub(r'\(\s*(?:eastern|east|e|west|w)\s*\)', ' ', name, flags=re.IGNORECASE)
-        name = re.sub(r'\b(?:eastern|east|west)\b', ' ', name, flags=re.IGNORECASE)
-
-        # Remove leading parenthetical prefixes
-        while name.lstrip().startswith('('):
-            new_name = re.sub(r'^\s*\([^\)]+\)\s*', '', name)
-            if new_name == name:
-                break
-            name = new_name
-
-        # Build pattern list based on flags
-        patterns_to_apply = []
-        if ignore_regional:
-            patterns_to_apply.extend(REGIONAL_PATTERNS)
-        if ignore_geographic:
-            patterns_to_apply.extend(GEOGRAPHIC_PATTERNS)
-        if ignore_misc and ignore_regional:
-            patterns_to_apply.extend(MISC_PATTERNS)
-
-        for pattern in patterns_to_apply:
-            name = re.sub(pattern, '', name, flags=re.IGNORECASE)
-
-        # Apply user-configured ignored tags
-        for tag in user_ignored_tags:
-            escaped_tag = re.escape(tag)
-            if '[' in tag or ']' in tag or '(' in tag or ')' in tag:
-                name = re.sub(escaped_tag + r'\s*', '', name, flags=re.IGNORECASE)
-            else:
-                if re.match(r'^\w+$', tag):
-                    name = re.sub(r'\b' + escaped_tag + r'\b', '', name, flags=re.IGNORECASE)
-                else:
-                    name = re.sub(escaped_tag + r'\s*', '', name, flags=re.IGNORECASE)
-
-        # Remove callsigns in parentheses
-        if ignore_regional:
-            name = re.sub(r'\([KW][A-Z]{3}(?:-(?:TV|CD|LP|DT|LD))?\)', '', name, flags=re.IGNORECASE)
-        else:
-            name = re.sub(r'\([KW](?!EST\)|ACIFIC\)|ENTRAL\)|OUNTAIN\)|TLANTIC\))[A-Z]{3}(?:-(?:TV|CD|LP|DT|LD))?\)', '', name, flags=re.IGNORECASE)
-
-        if ignore_regional:
-            name = re.sub(r'\([A-Z0-9]+\)', '', name)
-
-        # Remove common suffixes/prefixes
-        name = re.sub(r'^The\s+', '', name, flags=re.IGNORECASE)
-        name = re.sub(r'\s+Network\s*$', '', name, flags=re.IGNORECASE)
-        name = re.sub(r'\s+Channel\s*$', '', name, flags=re.IGNORECASE)
-        name = re.sub(r'\s+TV\s*$', '', name, flags=re.IGNORECASE)
-
-        # Strip decorative Unicode markers (◉, ², superscript letters ᴿᴬᵂ…)
-        # that some IPTV providers append as quality/status badges. Only
-        # characters in decorator categories are removed; accented letters
-        # (é, î, ü…) are in Ll/Lu and are preserved.
-        name = ''.join(
-            ' ' if unicodedata.category(c) in _DECORATOR_CATS else c
-            for c in name
-        )
-
-        # Clean up whitespace
-        name = re.sub(r'\s+', ' ', name).strip()
-
-        if not name:
-            self.logger.debug(f"normalize_name returned empty for: '{original_name}'")
-
-        return name
-
-    def calculate_similarity(self, str1, str2, min_ratio=0.0):
-        """Levenshtein distance-based similarity ratio (0.0 to 1.0).
-        If min_ratio > 0, returns 0.0 early when the result can't reach it."""
-        if len(str1) < len(str2):
-            str1, str2 = str2, str1
-        len1, len2 = len(str1), len(str2)
-        if len2 == 0 or len1 == 0:
-            return 0.0
-
-        total_len = len1 + len2
-        # Length-difference pre-check: even with 0 substitutions, the distance
-        # is at least (len1 - len2), so the max possible ratio is bounded.
-        if min_ratio > 0:
-            max_possible = (total_len - (len1 - len2)) / total_len
-            if max_possible < min_ratio:
-                return 0.0
-            # Max allowed distance to still meet min_ratio
-            max_distance = int(total_len * (1.0 - min_ratio))
-
-        previous_row = list(range(len2 + 1))
-        for i, c1 in enumerate(str1):
-            current_row = [i + 1]
-            for j, c2 in enumerate(str2):
-                insertions = previous_row[j + 1] + 1
-                deletions = current_row[j] + 1
-                substitutions = previous_row[j] + (c1 != c2)
-                current_row.append(min(insertions, deletions, substitutions))
-            # Early termination: if the minimum value in this row already
-            # exceeds max_distance, no subsequent row can produce a valid result
-            if min_ratio > 0 and min(current_row) > max_distance:
-                return 0.0
-            previous_row = current_row
-
-        distance = previous_row[-1]
-        return (total_len - distance) / total_len
-
-    @staticmethod
-    def _length_scaled_threshold(base_threshold, shorter_len):
-        """Require higher similarity for shorter strings to avoid false positives."""
-        if shorter_len <= 4:
-            return max(base_threshold, 95)
-        elif shorter_len <= 8:
-            return max(base_threshold, 90)
-        return base_threshold
-
-    @staticmethod
-    def _has_token_overlap(str_a, str_b, min_token_len=4, require_majority=False):
-        """Check that distinctive tokens are shared between two strings.
-
-        Basic mode: at least one token (>= min_token_len) must be shared.
-        Majority mode: uses all tokens (>= 2 chars) and requires that more than
-        half of the smaller set overlaps. Catches false positives like
-        "america racing" vs "america bbc" while allowing single-token matches.
-        """
-        # "network"/"channel"/"television" are generic brand-suffix words, not
-        # distinctive - treat as common so the subset guard does not reject
-        # cases like "FanDuel Sports Cincinnati" vs "FanDuel Sports Network
-        # Cincinnati". (They're already stripped from end-of-string by
-        # normalize_name; this catches mid-string occurrences.)
-        common_words = {
-            "the", "and", "of", "in", "on", "at", "to", "for", "a", "an",
-            "network", "channel", "television",
-        }
-
-        if require_majority:
-            # Use all meaningful tokens (>= 2 chars) for stricter checking.
-            # Single-digit tokens (1, 2, 3, ...) are kept because they're
-            # channel-distinguishing (BBC 1 vs BBC 2, ESPN 1 vs ESPN 2 etc.)
-            # even though they're only 1 char.
-            def _meaningful(t):
-                if t in common_words:
-                    return False
-                return len(t) >= 2 or t.isdigit()
-            tokens_a = {t for t in str_a.split() if _meaningful(t)}
-            tokens_b = {t for t in str_b.split() if _meaningful(t)}
-            if not tokens_a or not tokens_b:
-                return True
-            shared = tokens_a & tokens_b
-            if not shared:
-                return False
-            smaller = min(len(tokens_a), len(tokens_b))
-            if not len(shared) > smaller / 2:
-                return False
-
-            unique_a = tokens_a - tokens_b
-            unique_b = tokens_b - tokens_a
-
-            # Subset guard: when one side is a strict subset of the other and
-            # the larger side has a distinctive token the smaller lacks, the
-            # candidate is a more specific channel than the query and the high
-            # fuzzy score is a false positive. Catches e.g. "Nickelodeon" vs
-            # "Nickelodeon Teen". Threshold is >=4 chars; "live"/"now" are
-            # explicitly non-distinctive (stream label variants) so "ABC News"
-            # still matches "ABC News Live". Pure-digit tokens >=2 chars
-            # (e.g. "360") are also distinctive: "Canal+Sport" must not match
-            # "Canal+Sport 360".
-            if not unique_a:
-                if any(_is_distinctive(t) for t in unique_b):
-                    return False
-            elif not unique_b:
-                if any(_is_distinctive(t) for t in unique_a):
-                    return False
-
-            # Divergent guard: when BOTH sides have unique tokens AND at least
-            # one of those unique tokens is a distinctive (>=4 char) word, the
-            # strings describe different brands and the fuzzy score is
-            # misleading. Catches "Sky Cinema Disney" vs "Sky Cinema Decades"
-            # (decades = 7 chars), "Sky Cinema Fast" vs "Sky Cinema Family",
-            # and "BBC One vs BBC Two" once number-words become digits earlier.
-            # Number-word→digit normalization (in normalize_name) reduces this
-            # guard's risk: "Three Angels" / "3 Angels" both become "3 angels"
-            # before reaching here, so they never trigger this case.
-            if unique_a and unique_b:
-                if any(len(t) >= 4 for t in unique_a | unique_b):
-                    return False
-
-            # Numeric/ordinal divergent guard: when BOTH sides have a numeric
-            # or ordinal token unique to them, they are sibling channels
-            # distinguished by number (BBC One vs BBC Two; ESPN 1 vs ESPN 2).
-            # Short tokens like "one"/"two" wouldn't trip the divergent guard
-            # above so they need their own check.
-            _NUMERIC = {
-                "one", "two", "three", "four", "five", "six", "seven", "eight",
-                "nine", "ten", "eleven", "twelve",
-                "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12",
-                "first", "second", "third", "fourth", "fifth",
-            }
-            if (unique_a & _NUMERIC) and (unique_b & _NUMERIC):
-                return False
-
-            return True
-
-        # Basic mode: at least one long token shared
-        tokens_a = {t for t in str_a.split() if t not in common_words and len(t) >= min_token_len}
-        tokens_b = {t for t in str_b.split() if t not in common_words and len(t) >= min_token_len}
-        if not tokens_a or not tokens_b:
-            return True
-        return bool(tokens_a & tokens_b)
-
-    def process_string_for_matching(self, s):
-        """Normalize for token-sort matching: lowercase, remove accents, sort tokens."""
-        s = unicodedata.normalize('NFD', s)
-        s = ''.join(char for char in s if unicodedata.category(char) != 'Mn')
-        s = s.lower()
-        s = re.sub(r'([a-z])(\d)', r'\1 \2', s)
-        cleaned_s = ""
-        for char in s:
-            if 'a' <= char <= 'z' or '0' <= char <= '9':
-                cleaned_s += char
-            else:
-                cleaned_s += ' '
-        tokens = sorted([token for token in cleaned_s.split() if token])
-        return " ".join(tokens)
 
     @staticmethod
     def _is_group_header(name):
@@ -692,15 +360,6 @@ class FuzzyMatcher:
         is a common in-name separator ("001 | Team A vs Team B")."""
         return bool(re.search(r'[#=*]{2,}|\|{3,}', name or ""))
 
-    @staticmethod
-    def _trailing_number(name):
-        """Return the integer value of a space-separated, purely-numeric
-        trailing token, or None. 'HBO 2' -> 2, 'DIRECTV 4K Live 1' -> 1,
-        'ESPN' -> None, 'ESPN2' -> None (digit not space-separated). Used to
-        reject 'Foo 1' vs 'Foo 2' false positives -- different channels that
-        otherwise fuzzy-match almost perfectly."""
-        m = re.search(r'(?:^|\s)(\d{1,4})\s*$', name or "")
-        return int(m.group(1)) if m else None
 
     def _channel_number_boost(self, stream_name, expected_number):
         """
@@ -719,95 +378,6 @@ class FuzzyMatcher:
             return 5
         return 0
 
-    # Words that match the callsign regex shape but are never US broadcast
-    # callsigns. A US callsign (K/W + 2-4 letters) is shape-identical to many
-    # common English words, so the loose Priority-4 pattern mis-extracts them -
-    # e.g. "with" in "Bizarre Foods with Andrew Zimmern" becomes callsign
-    # "WITH". Regex alone cannot tell "WITH" from "WABC"; frequent K/W-initial
-    # words are denied explicitly so they never extract as a callsign. WWE/WWF/
-    # WCW stop wrestling show names extracting as false-positive callsigns.
-    _CALLSIGN_DENYLIST = frozenset({
-        'WWE', 'WWF', 'WCW', 'EAST',
-        'WAR', 'WARS', 'WARM', 'WASH', 'WATCH', 'WAVE', 'WAVES', 'WAY', 'WAYS',
-        'WEB', 'WEEK', 'WELL', 'WENT', 'WERE', 'WEST', 'WHAT', 'WHEN', 'WHERE',
-        'WHICH', 'WHILE', 'WHITE', 'WHO', 'WHY', 'WIDE', 'WIFE', 'WILD', 'WILL',
-        'WIND', 'WINE', 'WING', 'WINGS', 'WINS', 'WIRE', 'WISE', 'WISH', 'WITH',
-        'WOLF', 'WOMAN', 'WOMEN', 'WOOD', 'WORD', 'WORDS', 'WORK', 'WORKS',
-        'WORLD', 'WORM', 'WORN', 'WRAP',
-        'KEEN', 'KEEP', 'KEPT', 'KEY', 'KEYS', 'KICK', 'KID', 'KIDS', 'KILL',
-        'KIND', 'KING', 'KINGS', 'KISS', 'KITE', 'KNEE', 'KNEW', 'KNOW', 'KNOWN',
-    })
-
-    def _compute_callsign_with_confidence(self, channel_name):
-        """
-        Extract US TV callsign with a confidence flag.
-
-        Returns (callsign, is_high_confidence). High confidence =
-        Priorities 1-3 (parenthesized / suffixed-paren / end-of-name).
-        Priority 4 (any loose word) is low confidence. (None, False)
-        when nothing extractable.
-        """
-        # Remove common provider prefixes
-        channel_name = re.sub(r'^D\d+-', '', channel_name)
-        channel_name = re.sub(r'^USA?\s*[^a-zA-Z0-9]*\s*', '', channel_name, flags=re.IGNORECASE)
-
-        # Priority 1: Callsigns in parentheses (most reliable)
-        paren_match = re.search(r'\(([KW][A-Z]{3})(?:-[A-Z\s]+)?\)', channel_name, re.IGNORECASE)
-        if paren_match:
-            callsign = paren_match.group(1).upper()
-            if callsign not in self._CALLSIGN_DENYLIST:
-                return callsign, True
-
-        # Priority 2: Callsigns with suffix in parentheses
-        paren_suffix_match = re.search(r'\(([KW][A-Z]{2,4}-(?:TV|CD|LP|DT|LD))\)', channel_name, re.IGNORECASE)
-        if paren_suffix_match:
-            return paren_suffix_match.group(1).upper(), True
-
-        # Priority 3: Callsigns at the end
-        end_match = re.search(r'\b([KW][A-Z]{2,4}(?:-(?:TV|CD|LP|DT|LD))?)\s*(?:\.[a-z]+)?\s*$', channel_name, re.IGNORECASE)
-        if end_match:
-            callsign = end_match.group(1).upper()
-            if callsign not in self._CALLSIGN_DENYLIST:
-                return callsign, True
-
-        # Priority 4: Any word matching callsign pattern (low confidence)
-        word_match = re.search(r'\b([KW][A-Z]{2,4}(?:-(?:TV|CD|LP|DT|LD))?)\b', channel_name, re.IGNORECASE)
-        if word_match:
-            callsign = word_match.group(1).upper()
-            if callsign not in self._CALLSIGN_DENYLIST:
-                return callsign, False
-
-        return None, False
-
-    def _extract_callsign_with_confidence(self, channel_name):
-        """
-        Cached wrapper around _compute_callsign_with_confidence.
-
-        Extraction is pure in channel_name, so results are memoized - the
-        anchor calls this once per stream over a fixed candidate list, which
-        is otherwise massively redundant across the per-channel matching
-        loop. Cache is cleared by precompute_normalizations.
-        """
-        cached = self._callsign_cache.get(channel_name)
-        if cached is not None:
-            return cached
-        result = self._compute_callsign_with_confidence(channel_name)
-        self._callsign_cache[channel_name] = result
-        return result
-
-    def extract_callsign(self, channel_name):
-        """
-        Extract US TV callsign from channel name with priority order.
-        Returns None if common false positives appear alone.
-        """
-        callsign, _ = self._extract_callsign_with_confidence(channel_name)
-        return callsign
-
-    def normalize_callsign(self, callsign):
-        """Remove the broadcast suffix (-TV/-CD/-LP/-DT/-LD) from a callsign."""
-        if callsign:
-            callsign = re.sub(r'-(?:TV|CD|LP|DT|LD)$', '', callsign)
-        return callsign
 
     def alias_match(self, lineup_name, candidate_names, alias_map, user_ignored_tags=None):
         """
